ResumoExcel/requirements.txt
ipykernel==6.29.4
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.4
openai==1.24.0
pypdf==4.2.0
streamlit==1.33.0
python-dotenv==1.0.1
faiss-cpu==1.8.0
pandas==2.0.3

